third_party/WebKit/LayoutTests/fast/canvas/webgl/webgl-exceptions-expected.txt

PASS gl.getBufferParameter() threw exception TypeError: Failed to execute 'getBufferParameter' on 'WebGLRenderingContextBase': 2 arguments required, but only 0 present..
PASS gl.getRenderbufferParameter() threw exception TypeError: Failed to execute 'getRenderbufferParameter' on 'WebGLRenderingContextBase': 2 arguments required, but only 0 present..
PASS gl.getTexParameter() threw exception TypeError: Failed to execute 'getTexParameter' on 'WebGLRenderingContextBase': 2 arguments required, but only 0 present..
PASS gl.getVertexAttrib() threw exception TypeError: Failed to execute 'getVertexAttrib' on 'WebGLRenderingContextBase': 2 arguments required, but only 0 present..
PASS gl.getAttachedShaders() threw exception TypeError: Failed to execute 'getAttachedShaders' on 'WebGLRenderingContextBase': 1 argument required, but only 0 present..
PASS gl.getExtension() threw exception TypeError: Failed to execute 'getExtension' on 'WebGLRenderingContextBase': 1 argument required, but only 0 present..
PASS gl.getFramebufferAttachmentParameter() threw exception TypeError: Failed to execute 'getFramebufferAttachmentParameter' on 'WebGLRenderingContextBase': 3 arguments required, but only 0 present..
PASS gl.getParameter() threw exception TypeError: Failed to execute 'getParameter' on 'WebGLRenderingContextBase': 1 argument required, but only 0 present..
PASS gl.getProgramParameter() threw exception TypeError: Failed to execute 'getProgramParameter' on 'WebGLRenderingContextBase': 2 arguments required, but only 0 present..
PASS gl.getShaderParameter() threw exception TypeError: Failed to execute 'getShaderParameter' on 'WebGLRenderingContextBase': 2 arguments required, but only 0 present..
PASS gl.getUniform() threw exception TypeError: Failed to execute 'getUniform' on 'WebGLRenderingContextBase': 2 arguments required, but only 0 present..
PASS gl.uniform1iv() threw exception TypeError: Failed to execute 'uniform1iv' on 'WebGLRenderingContextBase': 2 arguments required, but only 0 present..
PASS gl.uniformMatrix2fv() threw exception TypeError: Failed to execute 'uniformMatrix2fv' on 'WebGLRenderingContextBase': 3 arguments required, but only 0 present..
PASS gl.uniform1fv() threw exception TypeError: Failed to execute 'uniform1fv' on 'WebGLRenderingContextBase': 2 arguments required, but only 0 present..
PASS gl.vertexAttrib1fv() threw exception TypeError: Failed to execute 'vertexAttrib1fv' on 'WebGLRenderingContextBase': 2 arguments required, but only 0 present..
PASS gl.uniformMatrix4fv(1, 1) threw exception TypeError: Failed to execute 'uniformMatrix4fv' on 'WebGLRenderingContextBase': 3 arguments required, but only 2 present..
PASS successfullyParsed is true

TEST COMPLETE
